+++ serialization-msgpack/src/commonTest/kotlin/com/ensarsarajcic/kotlinx/serialization/msgpack/extension/MsgPackExtensionDecoderTest.kt
@@ -26,6 +26,18 @@ internal class MsgPackExtensionDecoderTest {
     }
 
     @Test
+    fun testExtensionVariableDecode() {
+        val input = "c71a016c6a686673646c6a6668736b66687364686b6673646a66687364"
+        val expectedOutput = MsgPackExtension(0xc7.toByte(), 0x01, "6c6a686673646c6a6668736b66687364686b6673646a66687364".hexStringToByteArray())
+        val decoder = MsgPackDecoder(MsgPackConfiguration.default, SerializersModule {}, input.hexStringToByteArray().toMsgPackBuffer())
+        val serializer = MsgPackExtension.serializer()
+        val result = serializer.deserialize(decoder)
+        assertEquals(expectedOutput.type, result.type)
+        assertEquals(expectedOutput.extTypeId, result.extTypeId)
+        assertEquals(expectedOutput.data.toList(), result.data.toList())
+    }
+
+    @Test
     fun testCustomExtensionDecode() {
         val input = "d5030102"
         val expectedOutput = CustomExtensionType(listOf(0x01, 0x02))